launch-client(1): Describe communication outside containers, too
s-r-launch-client was originally written to insert debugging commands
into a per-game container, but it can equally well be used to escape
from a per-game container and run commands outside. This mechanism
should be used cautiously, because it relies on implementation details
of the host system, but it's sometimes necessary for development and
debugging.

Signed-off-by: Simon McVittie <smcv@collabora.com>