populate-depot: Always generate ./run
This was accidentally made conditional on whether we are generating a
tool manifest for the Steam compatibility tool framework, which we
don't want to do if we are using the container runtime for some other
purpose. For example, the container runtime that's used to run
`steamwebhelper` doesn't need the tool manifest.

Signed-off-by: Simon McVittie <smcv@collabora.com>