container-runtime: Don't fail testing if libgbm.so.1 is incomplete
We currently use Debian 10 and Steam Runtime 2 to smoke-test runtimes,
but those have an older version of libgbm.so.1 than the one that will
"naturally" be assumed by `steamwebhelper` when compiled on
Steam Runtime 3 'sniper'.

We unconditionally take libgbm.so.1 from the host system because it must
be kept in lockstep with the rest of Mesa, so any time we are checking
the ABI of libgbm.so.1 in detail, we will need to be prepared to ignore
certain missing symbols. In practice, `steamwebhelper` does not use this
specific symbol, so it's OK.

Signed-off-by: Simon McVittie <smcv@collabora.com>