tests: Ensure we have better coverage if built with --werror
Release builds are not done with --werror, so we will gracefully skip
various parts of the build if we are running on a particularly strange
platform (where /dev/pts isn't a directory, /dev/stderr isn't a symlink,
or we are unable to expand `${LIB}` and `${PLATFORM}`). But for our
official CI and individual developers' builds, we expect the build to
be done on a platform that makes sense, so we can be more strict.

Signed-off-by: Simon McVittie <smcv@collabora.com>