build-aux: Add convenience files to cross-compile for arm64
This is a useful way to check that non-x86 builds aren't regressing,
without actually needing a non-x86 device.

Not all tests will pass in this configuration because qemu-user
is not a perfect emulation of a real aarch64 machine (for example
PR_GET_CHILD_SUBREAPER is unimplemented), but it's better than nothing.
In particular, the adverb-preload and wrap-setup tests pass, and those
are the ones I'm altering for steamrt/tasks#595.

Signed-off-by: Simon McVittie <smcv@collabora.com>