pressure-vessel: Avoid looping through zero-length array of emulator hosts
If compiled for a non-x86 architecture, we don't expect to be emulated
by some foreign architecture with a mechanism like FEX or qemu-user,
so PV_N_SUPPORTED_ARCHITECTURES_AS_EMULATOR_HOST is zero.
In this case don't try to allocate and loop over an array of length 0.
This also avoids gcc warnings with -Wtype-limits,
"comparison of unsigned expression in ‘< 0’ is always false".

Signed-off-by: Simon McVittie <smcv@collabora.com>