tests: Combine expected test results for x86_64 and non-x86
There are three differences between our handling of x86 and non-x86 here:

1. On x86, x86_64 is always treated as the primary architecture,
   with i386 as secondary (even if pressure-vessel itself is compiled
   as an i386 binary). On other platforms, the native platform is
   primary.
2. On x86, i386 is available as a secondary architecture.
   On other platforms, there its no secondary architecture.
3. We have different mock expansions for `${LIB}`, `${PLATFORM}` and
   the multiarch tuple on x86 and non-x86.

By abstracting the "primary architecture", we can avoid having lines
of test code that are only run on non-x86 platforms, which in practice
means they are not routinely run at all. This makes them less likely
to regress.

Signed-off-by: Simon McVittie <smcv@collabora.com>