pv-wrap: Use a more realistic multiarch tuple for unit tests
All of our supported architectures define _SRT_MULTIARCH from the build
system, and pressure-vessel can't be compiled for other architectures
without defining _SRT_MULTIARCH, so we don't need to override this
to a mock string; we can leave it set to the real multiarch tuple of the
architecture.

This avoids setting up inconsistencies between this code path, and other
parts of pressure-vessel that do not force a mock architecture.

Signed-off-by: Simon McVittie <smcv@collabora.com>