pressure-vessel: Add STEAM_COMPAT_RUNTIME_SDL2=sdl2-compat
After we add sdl2-compat to container runtimes, we will presumably want
a reasonably long testing period in which the default is still
"classic" SDL2, but users, developers and testers can opt-in to using
sdl2-compat as a replacement for it. This mechanism lets us achieve that.

This will not be practically useful until soldier and/or sniper includes
a copy of sdl2-compat.

steamrt/tasks#579, steamrt/tasks#573

Signed-off-by: Simon McVittie <smcv@collabora.com>