launch-options: Allow control over SDL dynamic API overrides
By default, we will use a bundled copy of SDL 2 or SDL 3 if a game happens
to provide one: for example, Hitman (2016) has a bundled copy of SDL
2.0.5, and Dota 2 has a bundled copy of some intermediate version between
3.1.6 and 3.1.8 at the time of writing. When using pressure-vessel,
we can override that with the Steam Runtime's newer, presumably-better
version (steamrt/tasks#578).

For SDL 2, in runtime versions that contain sdl2-compat as a non-default
version, we can do the same to replace classic SDL 2 with sdl2-compat
(steamrt/tasks#579).

Expose this in steam-runtime-launch-options to make it more discoverable
and easier to test.

Signed-off-by: Simon McVittie <smcv@collabora.com>