wrap-setup test: Stop asserting that "/" is marked as exported
This makes no sense: the comment above indicates that we *didn't* want
this path to be exported, and if it was exported then we would be
unable to have any control over the filesystem namespace.

In fact, it is not exported: the fact that it appears to be exported
is an implementation quirk of flatpak_exports_path_get_mode(),
which returns FLATPAK_FILESYSTEM_MODE_READ_WRITE because none of the
zero remaining path segments is unmapped.

Signed-off-by: Simon McVittie <smcv@collabora.com>