wrap-setup test: Systematically check ancestors of LD_PRELOADs
In two cases we were asserting that the grandparent directory is not
exported, and the LD_PRELOAD module itself is, but making no assertion
about the parent directory of the LD_PRELOAD module.

To make modules like MangoHud work more reliably, in fact we want to
export the parent directory in most cases.

steamrt/tasks#595

Signed-off-by: Simon McVittie <smcv@collabora.com>