CI: Build packages on Debian 11 'bullseye' aarch64
This demonstrates the ability to build on an architecture that doesn't
have the biarch special cases that x86 does: if we can build for this,
then in principle we can build for anything. aarch64 is a particularly
convenient test target because it's available from various cloud
providers including Google, and the Raspberry Pi (v3 or later) is
inexpensive and widely available for developer use.

steamrt/tasks#669, steamrt/tasks#679

Signed-off-by: Simon McVittie <smcv@collabora.com>