build-relocatable-install: Discover library dependencies automatically
Instead of hard-coding the list of library packages that we depend on,
look at the output of capsule-capture-libs and find out which package
owns the library that each symlink points to. Those packages are our
dependencies.

This is necessary if we want to run build-relocatable-install on
anything other than scout, because some of the libraries we depend on
have changed between Debian releases: for example libgio-2.0.so.0 used
to be in libglib2.0-0 but is now in libglib2.0-0t64. Similarly, some of
the libraries have changed their dependencies: GLib used to depend on
libffi.so.6 and libpcre.so.3, but now depends on libffi.so.8 and
libpcre2-8.so.0.

Signed-off-by: Simon McVittie <smcv@collabora.com>