build-relocatable-install: Automatically collect executables' dependencies
Instead of hard-coding selected libraries that our executables depend
on, we can ask capsule-capture-libs to inspect them and parse their
dependencies automatically.

One benefit of this is that it allows us to build a relocatable
installation of pressure-vessel on any Debian release, and expect that
relocatable installation to be runnable on any distro where glibc is equal
to or newer than the chosen Debian release, reducing our dependence on
Steam Runtime 1 'scout'.

Another benefit is that when building on scout, we no longer need to
bundle libraries or source code for libpcre (because scout's GLib backport
statically links its vendored copy instead, since steamrt/tasks#112) or
libselinux and util-linux (because we compile scout's GLib with SELinux
and libmount integration disabled, also since steamrt/tasks#112).

Signed-off-by: Simon McVittie <smcv@collabora.com>