build: Fix the linker option to select RPATH in preference to RUNPATH
This is not a -z sub-option, so this never worked. Previously we didn't
notice this because official pressure-vessel releases are built on
Steam Runtime 1 'scout', which is so old that it can't generate RUNPATHs
anyway.

As described in the comment, we genuinely do want RPATHs and not
RUNPATHs for the relocatable installation.

Signed-off-by: Simon McVittie <smcv@collabora.com>