wrap-setup: Bind-mount in the AT-SPI accessibility bus
AT-SPI has its own private D-Bus message bus, the "accessibility bus".
This is used to let applications export their window contents to
assistive technologies such as screen readers, and to let assistive
technologies send events to the applications. This avoids flooding the
session bus with AT-SPI traffic, which can be rather frequent.

If we want `steamwebhelper` to be accessible to assistive technologies,
it needs to be able to communicate with that bus. In older AT-SPI it
was an abstract AF_UNIX socket, which passes through "naturally",
but in newer AT-SPI it has been changed to be a filesystem-backed
AF_UNIX socket in order to allow sandboxing, therefore the socket needs
to be bind-mounted into our container.

As with the session bus, we can adapt Flatpak's code for this: Flatpak
would always use a proxy in order to impose a security boundary, but
pressure-vessel is specifically not a security boundary, so we can
simplify considerably.

steamrt/tasks#699

Signed-off-by: Simon McVittie <smcv@collabora.com>