check-va-api: Pair calls to vaDestroySurfaces() with create_surfaces()
Each test creates new surfaces and puts their IDs in the array `surfaces`,
and previously we only destroyed the surfaces created by the last test.
The result is that if we ran more than one test, the surfaces created
by earlier tests would be leaked.

For example, if the VA-API implementation does not support H264 but
does support MPEG, we would have leaked the pair of surfaces that were
created in order to test H264, but we would correctly destroy the pair
of surfaces that were created in order to test MPEG. Now we destroy both.

Signed-off-by: Simon McVittie <smcv@collabora.com>