check-va-api: Add an option to do all checks
We cannot know whether we expect the VA-API implementation to support
H264, MPEG or only postprocessing, so when called from Steam's
diagnostic tools we only require one success, and stop after the
first success. However, during development and debugging it's helpful
to be able to try all of the possible code paths.

Signed-off-by: Simon McVittie <smcv@collabora.com>