tests: Acknowledge a known issue with the ncurses family
Debian 11 shipped with a snapshot of ncurses which added new internal
symbols to libtinfo, which are associated with a short-lived internal
symver but are referenced by libncurses.

Steam Runtime 3 'sniper' inherited this version, causing mismatches when
pressure-vessel picks up the libtinfo from the host system (usually
newer, for example Debian 12 or 13) but keeps the libncurses from the
Steam Runtime.

Conversely, when we run medic or steamrt5 on Debian 11, the internal
symbol in libtinfo confuses pressure-vessel into thinking that the host
libtinfo might be newer even though it is really older, and once again
we get an incompatibility.

This is unfortunate, but is known and is not a regression, so don't
fail our tests when we encounter it. This will allow us to start
to test more systematically across multiple host OSs.

steamrt/tasks#703

Signed-off-by: Simon McVittie <smcv@collabora.com>