graphics-drivers-openxr-1: Deduplicate inactive runtimes by json_path
Sometimes, an inactive runtime's json_path was already seen as
another's json_path. In that case, we don't really care about it, so
just skip it. Make sure to be careful of ordering in the process, since
we don't want to skip a runtime that is able to be active with one that
cannot.

Signed-off-by: Ryan Gonzalez <ryan.gonzalez@collabora.com>