populate-depot: Rename away more problematic filenames
We already had a mechanism to avoid potentially problematic filenames
in a "minimized" runtime, because the Steampipe CDN wants all
filenames to be case-insensitively unique and avoid characters that
are not allowed by Windows. The Steam bootstrapper mechanism has the
additional restriction that it doesn't like filenames with commas, like
`files/share/i18n/charmaps/ISO_8859-1,GL.gz`, so we should disguise
those too.

Since the renaming mechanism is already in use for a significant number
of filenames and seems to work fine in practice, let's future-proof this
by switching from a denylist to an allowlist: preserve filenames with
ASCII letters, digits and a very limited amount of punctuation, but
disguise anything outside that subset, notably commas and Unicode.
This will protect us from possible restrictions that we don't yet
know about.

The practical effect of this in soldier, sniper and steamrt3c is to rename
`files/share/i18n/charmaps/ISO_8859-1,GL.gz`, `files/bin/[` and
`files/share/ca-certificates/mozilla/NetLock_Arany_=Class_Gold=_Főtanúsítvány.crt`
out of the way.

When pressure-vessel starts up a runtime, it uses the manifest to create
a non-minimized version in `var/tmp-*`, which has the original filenames
(as well as restoring permissions, empty files and directories, and
symlinks, which we also do not distribute). As a result we will still
get `/usr/share/i18n/charmaps/ISO_8859-1,GL.gz` and so on in the
running container.

steamrt/tasks#719

Signed-off-by: Simon McVittie <smcv@collabora.com>