tests: Don't assert that two different trees have same mtime
Normally in this test we are directly comparing the source with the
destination, so we can assert that the mtime is equal: in the case
where we are creating a hard-link, it's literally the same inode, and
in the case where we are not, we use utimensat() to set the destination
file's mtime equal to the source's.

However, when we're exercising `PV_COPY_FLAGS_USRMERGE`, we are not
actually comparing the source with the destination. Instead, we create
the source and the expected result of carrying out the /usr merge,
do the /usr merge into the destination, and compare the expected result
with the actual destination. In general this will result in
modification times in the expected result differing from those in the
actual result, although in practice this failure mode is rare: because
we only compare modification times to 1-second resolution, the test
only fails if the time taken to create the source and the expected result
happens to cross the boundary between one second and the next.

steamrt/tasks#724

Signed-off-by: Simon McVittie <smcv@collabora.com>