build-relocatable-install: Include version number in copyright file names
The x86 build of pressure-vessel is still done in a scout environment
for maximum backwards-compatibility (although this will become
unnecessary in August, when Steam officially stops supporting glibc
older than 2.31) but the aarch64 build is done in a bullseye environment.
This means we'll have two different versions of each library: the
one from scout, and the one from bullseye.

steamrt/tasks#708

Signed-off-by: Simon McVittie <smcv@collabora.com>