runtime: Write original path to Vulkan layers, etc. into JSON manifests
When we load a Vulkan driver or layer, EGL driver or external platform,
OpenXR runtime or layer, or a similar object via a JSON manifest,
the name that is listed in the manifest and passed to dlopen()
influences how related files will be loaded.

For example, if the module loads another library using the `${ORIGIN}`
dynamic string token, the path to the other library is resolved based
on the name that appeared in the JSON manifest; or if gdb is asked to
load detached debug symbols via a `.gnu_debuglink` reference, the
debug symbols will be looked up relative to the name that appeared in
the JSON manifest.

Until now, we would write a path in `…/overrides/lib/…` into the JSON
manifest. This is a symbolic link emitted by `capsule-capture-libs`.
However, `…/overrides` is only populated with the symlinks emitted by
`capsule-capture-libs`, which means that related files such as the
`steamoverlayvulkanlayer.so.dbg` that is referenced by
`steamoverlayvulkanlayer.so` will not be found if we look for them in
that directory.

In practice we expect the symbolic link emitted by `capsule-capture-libs`
to point to an absolute path - either the same path that was used outside
the container if it is in `/home`, `/opt`, etc. (for example
`/home/me/.steam/root/ubuntu12_64/steamoverlayvulkanlayer.so`),
or a path from outside the container prefixed with `/run/host/` or
similar if it would otherwise collide with our runtime (for example
`/run/host/usr/lib/x86_64-linux-gnu/mangohud/libMangoHud.so`).

If we write those paths into the JSON manifest, then the module itself
or an external tool like gdb can easily use them to locate related files
relative to the module, for example
`/home/me/.steam/root/ubuntu12_64/steamoverlayvulkanlayer.so.dbg` or
`/run/host/usr/lib/x86_64-linux-gnu/mangohud/libMangoHud_opengl.so`.

steamrt/tasks#595, steamrt/tasks#736

Signed-off-by: Simon McVittie <smcv@collabora.com>