populate-depot: Assert that our output is Steampipe-friendly
Steampipe doesn't accept non-Windows-compatible filenames like
`libcap2_1:2.22-1ubuntu3+steamrt1.1.txt` (the `:` is not allowed).

We cannot do this when including a SDK sysroot, because we already know
that the SDK is not Steampipe-compatible, due to the inclusion of files
like `/var/lib/dpkg/info/base-files:amd64.list` in unminimized form.

steamrt/tasks#746

Signed-off-by: Simon McVittie <smcv@collabora.com>