relocatable-install: Avoid epoch in copyright filenames
We're assuming here that source package names are always OK, but Debian
package version numbers can contain an epoch marker (starts with `1:`)
which introduce characters into the filename that are not compatible
with Windows filesystems or Steampipe.

Debian packages' versions are designed to be unique even when the epoch
is omitted, and the epoch is conventionally excluded from filenames
(for example the .dsc files that this script also handles) in order to
make them compatible with Windows filesystems.

steamrt/tasks#746

Fixes: 17ba613b "build-relocatable-install: Include version number in copyright file names"
Signed-off-by: Simon McVittie <smcv@collabora.com>