populate-depot: Relax restrictions on ~
For the version of the relocatable pressure-vessel tarball that contains
full source code, we want to be able to accommodate .dsc files, which
refer to the other source code by filenames that can contain ~ for
prerelease versions (`1.2~beta3`). Similarly it'll be easiest to keep
track of correct copyright and licensing information if we preserve the
version number (without epoch) as-is.

~ is not actually special to Unix shells unless it is at the beginning
of a word, so we can relax our previous restriction here.

steamrt/tasks#746

Signed-off-by: Simon McVittie <smcv@collabora.com>