docs: Mention the versions of apt, dpkg, glibc, Python on host systems
Most of what we do in the Steam Runtime relates to packages inside our
container or `LD_LIBRARY_PATH` environment, but some of our components
need to run on the host system, especially `steam-launcher`. Keep track
of what features we can and cannot safely use in those components.

In particular, note that Steam will soon drop support for glibc older
than 2.31, after which we will be able to assume a somewhat more
recent operating system (2020 or later).

Signed-off-by: Simon McVittie <smcv@collabora.com>