On non-x86, only set filter_exclusive_priority to 0 for sniper
On x86, titles that do not specify a runtime are assumed to be
targeting scout (as implemented by either the legacy runtime or
SLR 1.0), and it would be wrong to run them under a newer runtime.

Similarly on non-x86, titles that do not specify a runtime should be
assumed to be targeting the oldest runtime that could conceivably
support them, which in practice is sniper (currently only for arm64).
It would still be wrong to run titles that do not specify a runtime
under steamrt4, which is not 100% compatible with sniper.

Signed-off-by: Simon McVittie <smcv@collabora.com>